backend/requirements.txt
asgiref
boto3
charset-normalizer
cryptography
defusedxml
diff-match-patch
dj-database-url
dj-rest-auth
Django
django-allauth
django-cors-headers
django-import-export
django-redis
django-storages
djangorestframework
djangorestframework-simplejwt
et-xmlfile
gunicorn
idna
MarkupPy
oauthlib
odfpy
openpyxl
Pillow
psycopg2-binary
pycparser
PyJWT
python3-openid
pytz
PyYAML
requests
requests-oauthlib
sqlparse
tablib
typing_extensions
urllib3
whitenoise
xlrd
xlwt
zappa